Refactors validators into their own file
